Version 3.1
= Version 3.1 – April 12, 2025 =

+   Rework plugin environment check - limit when checked (on activate, updates, or daily).
+   Due to new extension (event_scheduler) and external dependencies, version set to 3.1.0.
    +   Version 3.0.4 was not released.
+   Added `EAC_ALLOWED_WP_SCHEDULES` to limit intervals shown on admin screen.
+   Added `allowed_schedules` filter to filter out any unwanted schedules/intervals.
+   security_cors: Use daily cron (if scheduled) to get host IP addresses.
+   security_cors: Suppress scheme/host warning.

= ~~Version 3.0.4 – March 31, 2025~~ =

+   Tested with WordPress 6.8.
+   New event_scheduler (cron) extension.
    +   Intervals - Manage custom intervals (aka schedules).
    +   Events - Schedule WP Core or custom interval events.
    +   Tasks - Add tasks (actions) to scheduled events.
+   `createScheduledEvents()` and `removeScheduledEvents()` called on plugin install/update now do nothing.
+   Added `plugin_activated` and `plugin_deactivated` actions.
+   Session extension - wait for WP 'init' before setting cookie.
+   Visitor Id - wait for WP 'init' before setting cookie.
+   Removed obsolete `delete_option(...)` statements.
+   Optimized `forEachNetworkSite()` method.
+   Removed schedule to purge transients, WP will do it (delete_expired_transients).
+   Don't flush object cache (`wp_cache_flush()`) when using external cache.
+   Added eacDoojigger_log_info|notice|warning|error|debug|always actions.
+   Debugging extension enhancements.
    +   Added wp-cron debugging options.
    +   Added support for Queue Monitor.
    +   Combined settings to single switch option.
+   `explode_with_keys()` now accepts array of strings to explode.

+++ EarthAsylumConsulting/abstract_backend.class.php
@@ -10,7 +10,7 @@ use EarthAsylumConsulting\Helpers\wp_config_editor;
  * @package		{eac}Doojigger
  * @author		Kevin Burkholder <KBurkholder@EarthAsylum.com>
  * @copyright	Copyright (c) 2025 EarthAsylum Consulting <www.earthasylum.com>
- * @version		25.0327.1
+ * @version		25.0411.1
  * @link		https://eacDoojigger.earthasylum.com/
  * @see 		https://eacDoojigger.earthasylum.com/phpdoc/
  * @used-by		\EarthAsylumConsulting\abstract_context
@@ -163,7 +163,7 @@ abstract class abstract_backend extends abstract_core
 		// when upgrade completes (old version is active)
 		add_action( 'upgrader_process_complete',			array( $this, 'plugin_admin_upgraded'), 10, 2 );
 		// Register admin_init to check for new install or upgraded version of the plugin
-		add_action('admin_init',							array( $this, 'plugin_admin_installed') );
+		add_action( 'admin_init',							array( $this, 'plugin_admin_installed') );
 	}
 
 
@@ -289,14 +289,9 @@ abstract class abstract_backend extends abstract_core
 	{
 		if ( $hook_extra['type'] === 'plugin' && in_array($hook_extra['action'], ['install','update']) )
 		{
-			if ( ! array_key_exists('plugins',$hook_extra) || in_array($this->PLUGIN_SLUG, $hook_extra['plugins']) )
+			if ( array_key_exists('plugins',$hook_extra) && in_array($this->PLUGIN_SLUG, $hook_extra['plugins']) )
 			{
 				$this->delete_site_transient( self::PLUGIN_HEADER_TRANSIENT );
-				$this->forEachNetworkSite(function()
-					{
-						$this->delete_site_transient( self::PLUGIN_HEADER_TRANSIENT );
-					}
-				);
 			}
 		}
 	}
